pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swf-common-lib"
version = "0.0.10"
description = "A common library for SWF components."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-json-logger",
    "psycopg2-binary",
    "requests"
]

[project.urls]
"Repository" = "https://github.com/bnlnpps/swf-common-lib"
